+++ computenestsupplier-20210521/core/Models/GetServiceResponseBody.cs
@@ -481,6 +481,16 @@ namespace AlibabaCloud.SDK.ComputeNestSupplier20210521.Models
         [Validation(Required=false)]
         public string CommodityCode { get; set; }
 
+        [NameInMap("ComplianceMetadata")]
+        [Validation(Required=false)]
+        public GetServiceResponseBodyComplianceMetadata ComplianceMetadata { get; set; }
+        public class GetServiceResponseBodyComplianceMetadata : TeaModel {
+            [NameInMap("CompliancePacks")]
+            [Validation(Required=false)]
+            public List<string> CompliancePacks { get; set; }
+
+        }
+
         /// <summary>
         /// <para>The time when the service was created.</para>
         ///